pyproject.toml
[project]
name = "malbeot-diary"
version = "0.1.0"
description = "AI 말벗 일기장"
requires-python = ">=3.13,<3.14"

dependencies = [
    # Core
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.1",
    "python-multipart==0.0.12",
    "click>=8.0.0",
    # Database
    "sqlalchemy==2.0.36",
    "alembic==1.14.0",
    "aiosqlite==0.20.0",       # 로컬 개발용 (SQLite)
    # "asyncpg==0.30.0",       # 프로덕션 배포용 (PostgreSQL)
    # Auth / JWT
    "python-jose[cryptography]==3.3.0",
    "bcrypt==4.0.1",
    "passlib[bcrypt]==1.7.4",
    # Cache / Session Store
    "redis==5.2.1",
    # AI / GPT
    "openai==1.57.2",
    # Scheduler
    "apscheduler==3.10.4",
    # Web Push
    "pywebpush==2.3.0",
    # Env
    "python-dotenv==1.0.1",
    # HTTP Client
    "httpx==0.27.2",
    "pydantic-settings>=2.13.1",
    "greenlet>=3.3.2",
    "azure-cognitiveservices-speech>=1.48.2",
]

[dependency-groups]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes (미사용 변수/임포트 등)
    "I",   # isort (import 정렬)
    "B",   # flake8-bugbear (잠재적 버그)
    "UP",  # pyupgrade (최신 파이썬 문법 권장)
]
ignore = [
    "E501",  # line too long (line-length로 관리)
    "B008",  # FastAPI Depends 패턴 허용
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
